frontend/requirements.txt
Flask
Flask-Login==0.2.9
Flask-Sqlalchemy
Flask-Admin
Flask-WTF
Flask-Migrate
Flask-Security
gunicorn
